requirements.txt
paho.mqtt
configparser

